Add asynchronous clock crossing to increase data rates for parallel trace
Remove support for 32-bit trace because it's an unlikely scenario and would
now be bottlenecked by the FIQ.

The videokit demo now runs at 200 MHz (i. e. 800 MB/s for 16-bit trace)
and the icicle demo runs at 175 MHz (i. e. 700 MB/s for 16-bit trace).
This seems to be the upper limit where we still get acceptable data
eyes. A custom board with a dedicated trace port and no need for
adapters might be able to go even higher.

The drive strength setting in the constraint files is now at maximum,
and this does improve the data eyes compared to the default setting.